edit: Use a permalink for each script
This would allow us to version the software we use in this workflow without being affected by upstream changes until we want to bump the version. This design adds more maintenance to this workflow, but it also protects users against unexpected issues that are outside of their control.

Discussed in https://github.com/nextstrain/ebola/pull/6#discussion_r1048835183